change device status to NotAvailable if block mode pv is provisioned on it.
Continuous discovery changes the device status (Available or NotAvailable) when pv is provisioned on it.
In case of filesystem mode pv, the discovery checks the filesystem on the device and makes it NotAvailable.
For block mode pv, the discovery checks the bind mount of the device in /var/lib/kubelet/plugins/kubernetes.io~local-volume/volumeDevices/<pv-name>/<pod-uuid>
and makes the device as NotAvailable if a bind mount is present.

Signed-off-by: Santosh Pillai <sapillai@redhat.com>