manifests: fix LSO descriptions
The long description which is visible from LSO details via OperatorHub
is no longer taken from metadata->annotations->description but
spec->description instead.

Metadata is now used for the short description displayed on the "tile"
of the OperatorHub.

Change the descriptions so that we provide a reasonably short
description on the tile and use a longer description when user clicks
the tile for more detailed view.