add tests for DeleteReconciler
Adding a few test scenarios for current behavior of DeleteReconciler
before making changes to it. Tests demonstrate how PVs are handled
during deletion and assume UID usage in the PV annotation
(AnnProvisionedBy). The usage of UID is not correct because this is
obtained from Node and if the Node is deleted the UID changes. This
results in incorrect cleanup of PVs during Reconcile() because
the deleter code attempts to delete only those PVs whose annotation
exactly matches. The comparison is done between PV annotation and
current runtimeconfig.Name - his is the unstable name which contains
Node name and node UID.