refactor DeleteReconciler constructor
RuntimeConfig from sig-storage-local-static-provisioner/pkg/common
module contains a mix of values that are static and dynamic.

Static values should be set in constructor only and don't change for the
entire lifecycle of DeleteReconciler - for example cache, client...

Dynamic values could be changed and so they need to be reloaded during
reconcile using ReloadRuntimeConfig helper - for example values from
config map, node name...

Ideally those should be separated but that would require changing the
vendored package but still the code can be refactored so the constructor
is cleaner and RuntimeConfig is taken out of the constructor. This
results in more readable code and better test code since RuntimeConfig
is where most fakes are needed for testing.