fix TestListBlockDevices unit test
Golang 1.20 coverage profiling feature sends a warning to stdout if -coverprofile flag is used with go test and no GOCOVERDIR is set:

warning: GOCOVERDIR not set, no coverage data emitted

In combination with faking command execution with helperCommand() this warning is appended to expected stdout in our tests and causes them to fail.

Adding GOCOVERDIR to the helper command prevents the warning, however this is still far from ideal solution to faking command executions.